Created bag_processing package (#431)
Made a new bag_processing package and moved some scripts from graph_bag here and added documentation to the scripts and the bag_processing readme. Also added a script to convert all bags in a directory from bayer to grayscale.
