+++ tools/bag_processing/package.xml
@@ -1,0 +1,19 @@
+<package>
+  <name>bag_processing</name>
+  <version>1.0.0</version>
+  <description>
+    The bag processing package
+  </description>
+  <license>
+    Apache License, Version 2.0
+  </license>
+  <author email="astrobee-fsw@nx.arc.nasa.gov">
+    Astrobee Flight Software
+  </author>
+  <maintainer email="astrobee-fsw@nx.arc.nasa.gov">
+    Astrobee Flight Software
+  </maintainer>
+  <buildtool_depend>catkin</buildtool_depend>
+  <build_depend>rosbag</build_depend>
+  <run_depend>rosbag</run_depend>
+</package>
