+++ tools/bag_processing/scripts/apply_histogram_equalization_to_images.py
