+++ tools/readme.md
@@ -3,6 +3,8 @@
 This directory contains tools that are useful for debugging,
 but are not run on the robot.
 
+\subpage bag_processing
+
 \subpage calibration
 
 \subpage ekfbag
